fix: generateWords moved to grid-generator
