+++ src/features/game-play/index.ts
@@ -2,7 +2,6 @@ export { GamePlay } from "./ui/GamePlay";
 export { GameHint } from "./ui/GameHint";
 export { GameHelp } from "./ui/GameHelp";
 export { SelectControls } from "./ui/SelectControls";
-export { generateWords } from "./lib/generateWords";
 export { useHint } from "./lib/useHint";
 export { GameScreen } from "./ui/GameScreen";
 export { TopPanel } from "./ui/TopPanel";
