+++ src/features/grid-generator/index.ts
@@ -1,4 +1,4 @@
-export { generateGridLetters } from "./lib/generator";
+export { generateGridLetters, generateWords } from "./lib/generator";
 export { fillRandomLetters, getRandomDirection } from "./lib/helpers";
 export { tryPlaceWord } from "./lib/placement";
 export type { DirectionCounts } from "./model/types";
