+++ src/features/grid-generator/lib/generator.ts
@@ -1,6 +1,8 @@
-import { clamp } from "~/shared/utils";
+import { wordsLibrary } from "~/shared/data/words";
+import { clamp, shuffle } from "~/shared/utils";
 import { fillRandomLetters, getWeightedDirection } from "./helpers";
 import { tryPlaceWord } from "./placement";
+import type { Difficulty, WordsCategory } from "~/shared/types";
 import type { DirectionCounts } from "../model/types";
 
 type GenerateGridLettersProps = {
@@ -13,13 +15,47 @@ type GenerateGridLettersReturn = {
   placedWords: Set<string>;
 };
 
+/**
+ * Generates array with each word from words array placed at random position. (words can intersect)
+ * Remaining positions are filled with random letters.
+ *
+ * @param {GenerateGridLettersProps} options options.
+ * @param {GenerateGridLettersProps['size']} options.size size of the grid.
+ * @param {GenerateGridLettersProps['words']} options.words words to place.
+ *
+ * @example
+ * ```ts
+ * const words = ["cat", "rat"];
+ * const size = 3;
+ * const { letters, placedWords } = generateGridLetters({ words, size });
+ *
+ *  letters:
+ *  Array [
+ *    ["f", "d", "r"],
+ *    ["m", "k", "a"],
+ *    ["c", "a", "t"]
+ *  ]
+ *  placedWords:
+ *  Set {
+ *    "cat",
+ *    "dog"
+ *  }
+ * ```
+ *
+ * @returns {GenerateGridLettersReturn} generated grid letters.
+ */
 function generateGridLetters({
   size,
   words,
 }: Readonly<GenerateGridLettersProps>): GenerateGridLettersReturn {
   let letters: string[][] = [];
   const placedWords = new Set<string>();
-  console.log(`WORDS TO PLACE: ${words.flat()}`);
+
+  const logging = false;
+
+  if (logging) {
+    console.log(`WORDS TO PLACE: ${words.flat()}`);
+  }
 
   const directionCounts: DirectionCounts = {
     horizontal: 0,
@@ -53,9 +89,11 @@ function generateGridLetters({
         dir,
       });
       if (succeeded) {
-        console.log(
-          `PLACED ${word} AT ${row}, ${col} WITH DIRECTION {${dir.dr},${dir.dc}}`
-        );
+        if (logging) {
+          console.log(
+            `PLACED ${word} AT ${row}, ${col} WITH DIRECTION {${dir.dr},${dir.dc}}`
+          );
+        }
         letters = result;
         // Update direction count
         if (dir.dr === 0 && dir.dc === 1) directionCounts.horizontal++;
@@ -67,7 +105,7 @@ function generateGridLetters({
       attempts++;
     }
 
-    if (!placed) {
+    if (!placed && logging) {
       console.log(`FAILED TO PLACE ${word} after ${maxAttempts} attempts`);
     }
     i++;
@@ -81,5 +119,95 @@ function generateGridLetters({
   };
 }
 
-export { generateGridLetters };
+type WordCountRange = {
+  min: number;
+  max: number;
+};
+
+type WordLengthRange = {
+  min: number;
+  max: number;
+};
+
+type GenerateWordsProps = {
+  difficulty: Difficulty;
+  category?: WordsCategory;
+  ranges: {
+    count: WordCountRange;
+    length: WordLengthRange;
+  };
+};
+
+/**
+ * Generates array with words of given difficulty and size.
+ * Words taken from words library.
+ *
+ * @param {GenerateWordsProps} options options.
+ * @param {GenerateWordsProps['difficulty']} options.difficulty difficulty of the puzzle.
+ * @param {GenerateWordsProps['category']} options.category category of the words.
+ * @param {GenerateWordsProps['ranges']} options.ranges ranges for word count and length.
+ *
+ * @example
+ * ```ts
+ * const words = generateWords({
+ *   difficulty: "easy",
+ *   category: "animals",
+ *   ranges: {
+ *     count: { min: 3, max: 4 },
+ *     length: { min: 3, max: 4 },
+ *   },
+ * });
+ *
+ *  words:
+ *  [
+ *    "cat",
+ *    "dog",
+ *    "bird"
+ *  ]
+ * ```
+ *
+ * @returns {string[]} generated words.
+ */
+function generateWords({
+  difficulty,
+  category,
+  ranges,
+}: GenerateWordsProps): string[] {
+  const categories = Object.keys(wordsLibrary) as WordsCategory[];
+  const selectedCategory =
+    category || categories[Math.floor(Math.random() * categories.length)];
+
+  const categoryWords = wordsLibrary[selectedCategory];
+  const difficultyWords = categoryWords[difficulty];
+
+  if (!difficultyWords) {
+    throw new Error(
+      `No words found for category ${selectedCategory} and difficulty ${difficulty}`
+    );
+  }
+
+  // Filter words by length suitable for grid size
+  const suitableWords = Array.from(difficultyWords).filter(
+    (word: string) =>
+      word.length >= ranges.length.min && word.length <= ranges.length.max
+  );
+
+  if (suitableWords.length === 0) {
+    throw new Error(
+      `No suitable words found for size ${ranges.length.max} and difficulty ${difficulty}`
+    );
+  }
+
+  const { min: minWords, max: maxWords } = ranges.count;
+  const wordCount = Math.min(
+    Math.floor(Math.random() * (maxWords - minWords + 1)) + minWords,
+    suitableWords.length
+  );
+
+  const shuffled = shuffle(suitableWords);
+
+  return shuffled.slice(0, wordCount);
+}
+
+export { generateGridLetters, generateWords };
 export type { GenerateGridLettersProps, GenerateGridLettersReturn };
